requirements.txt
rdkit-pypi==2022.9.5
torch>=2.2
torch-geometric
xgboost
scikit-learn
transformers
numpy==1.26.4
pandas
pytest
matplotlib
joblib
deepchem